Rename anchor generator
