+++ src/otx/algo/detection/heads/anchor_generator.py
